requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
python-chess>=1.10.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0
jupyter>=1.0.0
matplotlib>=3.7.0
fastapi>=0.110.0
uvicorn>=0.27.0
neo4j>=5.0.0
flask>=3.0.0
flask-socketio>=5.3.0